ONIE template updates
These changes:
Update the documentation in the README file, to follow a more
standardized format for DUE documenation, and provide
working examples of various ways builds can be invoked.

post-install-config-sh.template now installs python-sphinx only
on Debian 9 based images, as most users won't need it.
It's there to do ONIE documenation builds, so if you want the
exact environment used for build and release, that's it.

It also checks for the addition of the oniebuild user, in case
the script is being run repeatedly to debug.

duebuild
Still a work in progress.
Updated the --cbuild, --default commands to be more ONIE-centric.

due-bashrc.template - formatting clean up