Build with Pandoc is optional.
Pandoc generates the man pages from the markdown files,
but takes up anywhere from 50-150MB of additional space
when installed.

This seems like a bit of an ask to generate a few Kilobytes,
(at best) of a file.

This commit changes the makefile to regenerate the docs/due.1
man page only if:

  make -f ./debian/control docs

Is run, AND pandoc happens to be installed. This assumes that
the average user will just be building the Deb and not doing
active development on DUE that would require updating documentation.

So:
 dpkg-buildpackage -uc -us will build the deb without issues.

Running:
 make -f ./debian/control docs
...to build the docs target will generate a new docs/due.1

As this is a Git 3.0 formatted project, subsequent
dpkg-buildpackage invocations will COMPLAIN about it being out of
date until it is checked in, which is what we want in the case
of a file that should be changing this infrequently.