Add menu based --stop for containers to management menu
As one might expect, this stops a running contianer.
Use with caution since adventurous users can stop other
user's containers, and since all DUE containers are invoked with -rm
this means the data in the container is gone when it is stopped.

This features a 5 second timeout to allow the user to
question their life choices, and it uses 'logger' to leave a note
in the system log in case there is any question about what happened.

Also, removeed --view-all flag that didn't really do anything except
mess up filtering suggestions.