Added --contianer-name option to change default name of container
This was prompted by a build automation case where I was using
a container run a cleanup on a build directory.

Changing the name allows a subsequent inspection of running
containers to show what invoked the container and why, should
any debug be needed.

Note that changing the name to not have 'due' in it will
get it filtered out of any running container selection code
unless --all is also passed.

This may or may not be desirable behavior.