Fix tagged image select
This fixes a bug where image tags weren't really considered when
choosing an image to run, and cleans up pattern matching for
image names that were substrings of other images.

If an image name to use is submitted in name:tag format, all
available images are searched for a match to exactly 'name'
(it has to start at start of line, and end with a space), and if
a ':tag' is present, the results are filtered again to only show
image names with that particular tag.