Fix armel container creation
This adds a fix to update the Dockerfile  used in image
creation to add a copy of qemu-arm-static _before_ Docker
tries to execute any commands in the container if the container
is for an architecture that is different from the host's.

Note some images may come with an amd64 qemu-static-armel
(which, apparently, was how this feature was originally
working at all) and this step won't be necessary for them,
but it shouldn't hurt anything either.

Added:
	An armel example to the debian-package README.md
	Shellcheck cleanup,  which is what got me looking
	 at this in the first place.