Limit running containers per user
As people may forget they have containers running, this
introduces two features to manage that

1 - DUE_USER_CONTAINER_LIMIT, configurable in due.conf
Since due runs containers with a username, if the user
has more than this many containers running, launch will
fail until they are below this limit.

--run also prints the current number of running containers.

2 - --manage --stop now takes a pattern match argument, in
addition to supporting an interactive menu.
If a pattern is passed, it will generate a script of a list
of contaners to stop, in the same way --manage --delete-matched
does.

With a high number of containers, it's easier to edit a script
and fire away than to do individual stops. Also less error prone
as the user will see exactly what they are about to do.