Added --dockerarg to pass arguments to Docker invocation.
This also adds the ability (support may be too strong a word)
to run containers in privileged mode. While these changes
don't make it convenient, they do make it possible, and
given the need to "know what you're doing" when using these
options, this is probably a fair compromise.

This also has a change to container-create-user where it will
parse --docker-group-id to get the group ID of the host's
docker group. If the container's docker group ID does not
match the hosts, the container cannot start any other containers, so
this gets set up on login.

The FAQ was updated with an example and some notes on the dangers
of running continers with --privileged.