Added container specific run arguments section, fixed --tag
--tag wasn't getting passed to container creation correctly.

This is a customization hook to have containers automatically
get run with additional arguments.

Uses would be to automatically mount a host directory
(say, a library or cache of build tools), or to have the container
run in --priviliged mode (which I don't usually recommend, and
it's a little more difficult here for a reason)

As this feature becomes really useful in local customization, I
don't anticipate it having code that  would get pushed back upstream,
so it is left at the end of the file for cleaner merging on
DUE updates.