Debian package duebuild script uses --build-dsc for Debian source.
This refines the duebuild script's behavior when building
Debian packages from Debian Source Control (.dsc) files.

It now takes the name of a *.dsc file to build, and will attempt
a build with the information from that file.

Previously it tried to get too clever finding *.dsc and compressed
files, and that failed in the case of multiple *.dsc files
being present, and if the tar file noted its compression with
xz rather than .gz

This replaces the prior option --build-from-deb-source.

It is expected that users will either be building from code checked
out of source control, which already has a directory structure,
or they'll be downloading source from Debian, which comes with a
.dsc and tar.?z files.