Fix a bug where installing all dependencies is treated as failure.
Finally managed to run this down after seeing it off and on for months.
If there's a local package deps file around, it will fail to install,
even though all downloaded build dependencies do install.

This fix changes the logic so that if all packages listed as
build dependencies do install, the error condition is removed and
the build proceeds.