--use-local-repo adds initial support for a local Debian package repository.
This is helpful when building packages that depend on other
packages to have already been built.

Example - if package B depends on files supplied by package A
 cd package A dir
 duebuild --use-local-repo myRepo --cbuild
 (puts package A debs in myRepo)
 cd package B dir
 ( pulls package A debs from myRepo)
 duebuild --use-local-repo myRepo --cbuild
 (puts package B debs in myRepo)

In addition to creating the repository, the script:
1 puts an entry under /etc/apt/sources.list.d so that
  the build container will reference the repository.
2 Adds an /etc/apt/preferences.d file to set the local
  repository to have a higher priority than anything else.
  This assumes that, given a choice between a package
  coming from upstream, or what was built and put in the
  local repository, the local repository wins.

As the repository is created one directory level above the
build area, but DUE is smart enough to mount build containers
one directory up, the repository should persist on the host
system, and be available for subsequent invocations.

Note: only tested so far with amd64 and debs.
Adding source or multiple architectures should be tested
when a test case is available.