Debian package build: always apt-get upgrade
Running apt-get upgrade by default on a Debian package build
accomplishes two things:

1 - Assures that any package build dependencies that were already installed,
and meet the package build criteria, but ARE NOT the latest version will
get updated.

2 - Follows the behavior of Debian 'sbuild'

To elaborate. Apt-get upgrade will update packages that are NOT on in
the container, but if, say, an older version of GCC is present, that
will not get updated.
This update keeps build results consistent between containers of different
ages, at the cost of upgrading packages that may not impact the build.

It is felt that the additional time/resource hit introduced by doing the
upgrade is justified to avoid the subtle problems that could be introduced
if this doesn't happen.