Debian package build template: default build jobs are settable. Added --duebuild option.
This adds two changes, driven by me being some combination of lazy and finding
the existing interface a bit awkward.

First, specifying --jobs X or -jX on the command line won't have duebuild interpret
that as the start of a custom job, so in this case, the default of:
dpkg-buildpackage -us -uc -j4
(Invoked when a user runs due --build and selects the Debian package build
container with duebuild in it )
Can also be run as:
due --build --jobs 16
Resulting in:
dpkg-buildpackage -us -uc -j16

This is a user interface change that feels like a less awkward way to achive that.

Second, I added an alias for --build that is --duebuild to indicate that it
is running the /usr/local/bin/duebuild script in the container.
If --help is used after --duebuild, it will prompt the user for a container and
run 'duebuild --help' in the container to return a list of options there.
So rather than the user having to type:
  due --run --command /usr/local/bin/duebuild --help
They can use
  due --duebuild --help
...and get the same result.

Importantly, this does not change the behavior of due --build --help,
which prints the options that apply to how the build will run the container.

Also updated documentation and command line help to mention this feature.