--help is less order specific
I was getting annoyed with having to type an action argument
( --run, --manage, --create ) and then put --help or
--help-examples after it. While that makes sense from a hierarchical
point of view, a user looking for help might also start with
just needing help and then get more specific from there.

Which is a reasonable use case.

so due --help-examples --run
OR
due --run --help-examples

Will both run the examples function for run.

This also buffs out support for running a container's duebuild
script to get --help or --help-examples out of it.

So
due --duebuild --help-examples
OR
due --help-examples --duebuild
Will bring up the image menu to let the user choose the image
to run, and then it will get the --help-examples from the
duebuild script in that image.