Sanity check mounting multiple host directories to the same container directory
This comes up in particular if the user has their home directory mounted
elsewhere (defined in due.conf) and tries to build under ~/

Although it's also possible to hit this by adding bad --volume mount
instructions.

If the user's homedir was remapped, it was either convenience, or
something like the homedir is nfs mounted over multiple machines, and
since Docker mounting nfs can create weird permissions issues that don't
manifest as permissions issues...the homedir remap is a workaround.

Point being, if there's a remap, and the user tries to build under
~/   - flag it.
And if multiple host volumes try to mount to the same direcotry in
the container - flag that too.

Ultimately Docker will fail this, but wrapping it can be a bit more
user friendly - especially given the homedir redirect happens
behind the scenes and is less obvious.