Container export, image import works. --snapshot is --export
This was a long needed sanity check of exporting running containers
as images and verifying that they came back in as images.

The trick here for export was 'commit' to save all layers (I think)
of a running container - otherwise what imported back in failed to
find essential things like bash.

So, for --export, these Docker commands are used:
 commit - to save container as a local image
 save   - to export the image as tarfile
 rmi    - to delete the saved image (tarfile still exists)

For -- import it's more predicatble, with the docker command being:
 load

And duebuild scripts will take --build as a synonym for --cbuild.
This is intentionally left undocumented for now.