Use ncores to maximize build jobs.
Default to making every CPU core on the system available for
build, using ncores to get that number.

If ncores is not available, default to 4.

The user can still limit the number of jobs using --jobs, should
there be cases of a race condition where it is necessary
to build monothreaded.

Closes: #8 nproc