Rework saving containers, import/export
After giving it some thought, there are really three operations here:

1 - export an image as a file.
2 - import an image as a file.
3 - save a running container as an image.

So to address this, there are the following options under the
--manage menu

--export-image - select from existing images (or supply a name) and
 that image will be saved as a tarfile on disk, available to
 export to other systems.

--import-image - the reverse of export - take an image tarfile and
  import it to be runnable.

--export-container  is a bit different, as it affects a running
   CONTAINER and turns it into a local image on disk.

 This operation:
    1 - saves the container as a local image
	2 - exports that image as a tarfile on disk
	3 - deletes the local image so there is no entry to be run.

To run the container the user must use --import to add it.

The thinking with the delettion is that because it is _export_ container,
there is no local saving implied, and these images would accumuluate
over time. If the user really wants to save it, they can --import it.

We'll see how this pans out.