Add --manage --browse option
This is a first pass/experimental feature for dealing with
adding and pulling Docker images from Docker registries.

At the moment it is more of a reminder of the syntax involved in
browsing, pushing and pulling images, and will print examples
to do that for a given registry.
I tend to do this infrequently enough that I forget the syntax,
but not infrequently enough that I never need to do it.

This also adds curl (to browse the repository) and jq
(to format JSON data ) as requirements.