DUE_VERSION in libdue is the version source of truth.
The release version of DUE comes up in a number of places, like
the man pages and ChangeLog. With this change, the Makefile
can be used to set the version correctly in the changelog, and
in manpage-due-1.md, which Pandoc uses to create the DUE man page.

This also rolls in a documentation update after running spell
check on the manpage-due-1.md so that the man page is updated.