Update duebuild documentation. Add duebuild-readme.md files.
Started adding readme files for the duebuild scripts that come
with a container. It's mostly a print out of the command
line help, but easier to browse as it's a file outside of a
built container.

The command line help for the Debian package build duebuild
has been updated to be a bit clearer, with examples of the
differences between different build invocations, and the
common-packages example duebuild script has been updated to better
represent what should be commonly supported commands.