Debian-package duebuild - local repo path support, package update robustness
This adds the option to specify an absolute path for a local package
repository ( used if built packages are needed for subsequent builds
of other packages ).

If a the --use-local-repo argument starts with a / it is interpreted
as an absolute path in the container, and the package repository
will be placed accordingly. This is useful if not all packages are
being built at the same level in the build directory, as the
default action is to create the repository one directory up from the build
directory.

Also added changes to:
 * Keep the current package configuration on the upgrade of installed
packages
 * Not error out if no dependencies have to be installed.