More local repo support, build only binaries by default.
This changes the default duebuild behavior to NOT build source packages by default.
Dpkg-source throws an error when trying to build source out of a freshly apt-downloaded
source directory, and, since the point of a default configuration is to build something
reliably, and any users interested in building source are expected to have to debug
things it, not building the source seems to be a more sensible path. Users, of course
still have the option to supply additional arguments to build source.
This change was a necessary to test the due-manage-local-repo.sh script, below.

This also introduces due-manage-local-repo.sh to support further management of a
local package repository. Previously the --use-local-repo option would just
add the results of the current build to a repository, where they could be accessed
by another package. The due-manage-local-repo.sh script lets users add and remove
any .deb, and add or remove it from the local /etc/apt/sources.list.d/ directory.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>