Red Hat Linux compatibility - first steps.
This introduces some changes to support running DUE on a Red Hat Linux
system. Note that creating DUE compatible Red Hat images is not supported,
as there are a number of runtime differences to be sorted out, but
this should allow for running Debian containers on a Red Hat system.
Mostly this is around the Red Hat package manager dnf rather than apt.

A Red Hat user can:
 - install DUE with 'make install'
 - create any of the existing Debian based containers,
 - emulation through QEMU mostly works, (although I ended up having
   to reboot the system to get new qemu packages to exist in binfmt_misc.
   I'm probably missing something there.)

The Makefile install has been upgraded, and there are some additional
sanity checks for building containers that would be run by QEMU, as
things seem to work a bit differently.

There is also an issue where mounting the user's home directory in
the container required the container to be run --privileged. I'm not
a fan of this, but I haven't found a workaround for it yet, and without that,
everything else kinda doesn't work.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>