ChangeLog Lintian cleanup
Fixed a warning on the date
Also decided that the merge from master -> debian/master, which always
hits a snag when it discovers ChangeLog and Makefile do not exist on
the Debian branch and requries a manual 'git rm' of the files...
can continue to work the way it does.
The Debian branch could map ChangeLog to debian/changelog, but that
should probably be a manual merge to avoid wiping out any
Debian specific modifications.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>