Add local packages to container with  pre|post-install-local directories.
If a template has a pre-install-local directory, any packages in it
will be installed during the pre-install-conf.sh script running.

Likewise, if there is a post-install-local directory, any packages there
will be added when the post-install-conf.sh script runs.

This provides the option to build packages in to the container when
those packages cannot be pulled from a repository.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>