debian-package post-install-config missing DEBIAN_FRONTEND
A test of the local package install code failed because it
was missing DEBIAN_FRONTEND in front of 'noninteractive'.
This only came up if one was trying to include a package from
the build directory, rather than the more common case of
pulling the package from a Debian package repository.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>