debian-package/README.md - Fix Ubuntu 20.04 example
This sets the prompt in the example for creating an Ubuntu 20.04
container to be PKGU2004, rather than PKG1804.

This doesn't change that the created container would still be
Ubuntu 20.04 based, but does make it clear to the user that it
is not an Ubuntu 18.04 container when they log in.

Signed-off by: Alex Doyle <adoyle@nvidia.com>