debian-package template: Provide post-install-local directory
This updates the post-install-local.sh.template to handle a
post-install-local directory that has no *.deb files in it.

Rather than assume that the existence of the directory means
that there is something to install, it (currently)  checks the
directory for any *.deb files, and will install them if present.

This change in logic allows for the directory to be available
on checkout, and hold a README file that explains the purpose
of the directory, allowing users to be aware of this option
without having to find it by reading the documentation.

(Since we all know reading the documentation is a last resort :) )

Signed-off-by:  Alex Doyle <adoyle@nvidia.com>