debian-template: due-manage-local-repo.sh support mutliple local repos.
This updates due-manage-local-repo.sh to use the name of the repository
as the Codename, and to include it in the:
 /etc/apt/sources.list.d/due-local-<name>
 /etc/apt/preferences/due-local-<name>
files.

With this, named repositories can be enabled and disabled.

The delete repository code was also updated to try and clean
up the sources.list and preferences files if the directory
holding the repository had been deleted.

For those wondering why you would need more than one
local repository, I was testing a build that used
due-manage-local-repo.sh for a local package, and needed
a second, separate repository to act as my local Debian
package mirror to avoid surprising fellow devleopers
with suddenly experimental code.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>