libdue - qemu install is now only qemu-user-static
QEMU is only needed if the user is running Docker images
of an architecture that is different from the host.
Since it takes up a reasonable amount of storage space,
it is a 'Recommends' in Debian install, and
an install via 'make install' does not add it, since, in that case,
the lack of qemu will be noticed at run time, and installation
instructions will be supplied then.

Also, qemu-user-system is not required to run images of different
architectures as qemu-user-static supplies everything DUE is using.

This was flagged from Michael Tokarev who was cleaning up the
qemu metapackage (which is going away) and why only qemu-user-static
might be the only packge required.

It is refreenced as:
Bug 992947 - due wronly recommends qemu package.

Adding 'fixes' rather than 'closed' as the debian/master branch needs
it's debian/control file's dependency list updated.

Fixes: #992947

Signed-off-by: Alex Doyle <adoyle@nvidia.net>