libdue - support sub-type template directory file hierarchy
This allows for merging files from multiple sub directories
together with the ./templates/common-templates files to
use for image creation.

This is excessive unless one has a lot of containers that
have slight differences, which results in a lot of file
replication, or relative linking to common files.

For example, supporting three software version releases
with internal development and 'official release' tracks,
could require six different sources.list files and apt keys
to reference different software mirrors.

Rather than have each template soft link to common files,
the common files can be present in parent directories with
the unique files stored in sub directories.

When the file merge is done to create the directory the image
will be built out of, parent directories are merged first,
bottom most child directories last so that they may overwrite any
common files as needed.

(And if they need to delete parent files, that were not overwritten,
that's what the child's post-install script is for )

Each directory in the path named 'sub-type' indicates a new
 group of files to be copied in. So:

  due/templates/foo/sub-type/bar/sub-type/baz
   Results in a final merged directory of files that are:
    common-templates+(files from foo)+(files from bar)+(files from baz)
   ..written in that order, so that files from 'baz' will replace
    any identically named files provided by bar, foo, or common-templates.

Also updated templates/README.md and man page

Signed-off-by: Alex Doyle <adoyle@nvidia.com>