Makefile - ignore 'stash pop' error if nothing to pop.
When building the debian-package target, the debian/master
branch is checked out, so a 'git stash' is performed to save
any development changes, with a 'git pop' used to restore
the state of files at the end of the build.

If there are no changes, then the git pop will throw a
Makefile error.

These changes clean up the handling of what is:
"could not restore changes that were not there to begin with"

By using - to make the error non-fatal, and ||: to keep
it from printing out a message that says 'errors were ignored',
because that is not very comforting either.

Bottom line:
No need to scare people with an error that doesn't matter.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>