Support custom /etc/hosts in container
This could be needed if the run time environment
has DNS issues and hostname/ip address correlations
need to be supplied.

This is a bit more complicated than just adding a
./filesystem/etc/hosts file.

First - code has been updated to install the user supplied
/etc/hosts _before_ doing post-install updates, so an image
build will use the /etc/hosts.

Second - Docker replaces the /etc/hosts at run time, so
the container-create-user.sh script now looks for a
 /due-configuration/filesystem/etc/hosts
file, (which that would indicate it was a custom install )
and if it finds one, it replaces the existing /etc/hosts

This approach may need some refinement, as it may make
too many assumptions about what the run time should be.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>