/etc/hosts in container - append rather than replace
This changes the /etc/hosts replacement behavior to append
the contents of the /etc/hosts file supplied by the DUE template
so as to not overwrite the IP address Docker supplies to
the container at run time.

If a total replacement of /etc/hosts is required, the
container-create-user.sh script can be modified in the
template directory and will replace the default one
supplied by common-templates.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>