container-create-user.sh - append container /etc/hosts entries.
At run time, Docker provides an /etc/hosts file, so if one was
included as part of container configuration, it gets overwritten.

There are cases where the dynamic information supplied by Docker
should be kept, hence this change to append the image supplied
/etc/hosts data to the /etc/hosts that is present at run time.

Hopefully the combination creates the best of both worlds, rather
than a compromise that satisfies nobody.

Signed-off-by: Alex Doyle <alexddoyle@gmail.com>