Podman - initial support
This set of changes should detect if Podman is installed,
rather than Docker, and, if so, it will be used.

The assumption is that both are not installed on the system, as
Docker, expected to be the more common program, is the default,
and as such there is no way to choose one or the other.
(At least until someone comes up with a viable use case and
a pull request.)

There are some conditional cases for parsing Podman output, and
for reference, feature development was done using the Podman
version from Debian 11, should different versions introduce issues.

Signed-off-by: Alex Doyle <alexddoyle@gmail.com>