Makefile - Disable tarfile extraction for debian-package branch
At one point in packaging development it was useful to have the
original tarball extract in to a directory of just Debian specific
files.

However since the debian/master branch is a snapshot of what has
been upstreamed to Debian (version 3.0.0 as of this commit) it
should remain unchanged until the next upstream release.

This comments out the tarball extraction for now, so that it may
be reused when preparing another upstream release.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>