container-create-user.sh - Allow different named groups with the same ID
Use --non-unique with groupadd so if there is another group with the
same ID, the new name will be created with the same ID in /etc/groups.
The result is that there will be two entries in /etc/groups with
the same ID, but different names.
The goal is to have host-consistent groups on the files when
the contiainer exits, and to have the correct ID associated
with _either_ group name inside the container.

This should also reduce the chances of failure due to
host system or Docker image configuration, which the end user
may not be able to do much about.

Fortunately these collisions seem to be rare, as this is the
first time I've encountered it in 2 years of use.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>