due - --home-dir option now overrides the default.
This came up when I wanted to switch between home
directories for the same user on different filesystems.
Ex:
  /home/myUsername  <-default home directory
  /mnt/driveOne/home/myUsername <-home on a bigger drive
  /mnt/driveTwo/home/myUsername <-home on another bigger drive

So 'due --run --home-dir /mnt/driveOne/home/myUsername' should
have (and now does) set the home directory in DUE to
/mnt/driveOne/home/myUsername, rather than the default of the
host system's /home/myUsername.

This was due to the read of the value from ~/.conf/due/due.config
being read in and resetting the directory specified by --home-dir.

This fix checks to see that if the user has already specified a
home directory, then there is no need to pull in defaults.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>