Make --home-dir options more robust
This fixes and verifies directory mounting behavior
when the --home-dir has been used to change the user's
home directory on the comand line. Example of a
valid use case for this follows:

Background
------------
When mounting host directories, due will:
 Always mount the user's home directory
 Mount the current working directory (assumed to be under the user's home dir)
  ...unless it IS ALSO the user's home directory, then DON'T mount twice.
 Mount any additional directories specified by:
  --mount-dir <host path>:<container path>

Up to this point, DUE assumes the home directory is
 Under the user's specified home directory, which can be set by default using:
  Specified in /etc/due/due.conf
  Specified in ~/.conf/due/due.conf

This can hit issues if the user sets their home directory using:
  --home-dir < dir > on the command line after --run/--build

As the CWD and --home-dir may try to mount different home directories in the
container.

Example - a build system with two hard drives.
----------------------------------------------

 One is the host boot drive which has limited space,
 The other is mounted at /work/ssd has more build room.
 I normally want to build on the ssd, so my ~/.conf/due/due.conf
 file defaults my home directory in a container to map to:
   DUE_ENV_DEFAULT_HOMEDIR=/work/ssd

Both directories - ~/ and /work/ssd have a build directory 'foo'
For demonstration purposes, foo is Debian package source, which puts
the build products in the directory ABOVE the build directory.
DUE gets clever with this and will mount one directory up so that the
debs are save on the mounted host filesystem rather than in the container,
which will take the debs with it when it exits...

If I try to use DUE to build in directory 'foo' in my home directory
on the host drive, a due --build will fail because the following
volume mounts are passed to Docker:

 --volume /work/ssd:/home/adoyle     <- home
 --volume /home/adoyle:/home/adoyle  <- CWD

...and we can't mount two file systems at the same location in the container.

Ways to resolve this
--------------------

1 - Change my DUE home directory to the host drive
	./due --build --home-dir /home/adoyle

	...and then /home/adoyle/foo is under /home/adoyle, so the container
	will find it.

2 - have the other build mount it. Make sure you get your paths right,
    especially if it's a Debian package build that puts build products
	in the directory ABOVE the one you are building in.

	./due --run --mount-dir /home/adoyle/:/build --command cd /build/foo \; duebuild --build

 Will mount:
     --volume /home/adoyle/foo/:/build\
     --volume /work/ssd:/home/adoyle\
  ...and the build products end up under /home/adoyle

3 - Move the code under /work/ssd/foo
 And then cd /work/ssd/foo ; due --build
 will build locally as Docker mounts
     --volume /work/ssd:/home/adoyle  <- home dir
     --volume /work/ssd/foo:/work/ssd/foo <- unless it is a Debian package build,
	           in which case DUE gets clever and mounts /work/ssd at /work/ssd
			   so that the debs will be saved on the host file system.

Testing:
----------
Using a debian packge build for foo, with the above drives
Under default homedir on /work/ssd, in /work/ssd/foo

1 )
 Test: make sure builds still work.

 cd /work/ssd/foo

 Run: due --build

 Mounts:
     --volume /work/ssd:/home/adoyle\
     --volume /work/ssd:/work/ssd\
 Build products are in
     /work/ssd/foo.deb

This ^^ proves build works as it always has.

Delete /work/ssd/foo for the next test. Leave a copy at /home/adoyle/foo

2 )
 Test: Build in /home/adoyle/foo

 Run: due --build

 Mounts:
     --volume /work/ssd:/home/adoyle
 FAILS as /home/adoyle/foo is not accessible. The container wants /work/ssd/foo.

3 )
 Test: change DUE home directory mount to put /home/adoyle/foo under it.

 Run: due --build --home-dir /home/adoyle

 Mounts:
     --volume /home/adoyle:/home/adoyle
 Builds in /home/adoyle/foo, puts debs in /home/adoyle

4 )
 Test: Keep home dir at /work/ssd, build in /home/adoyle/foo as a mounted directory.
  (note the \; to keep the shell from evaluating the ; before the command gets parsed by DUE.

	./due --run --mount-dir /home/adoyle/:/build --command cd /build/foo \; duebuild --build

Mounts:
     --volume /home/adoyle/:/build\
     --volume /work/ssd:/home/adoyle\

 Builds in /home/adoyle/foo (or /build/foo from the container's perspective).

 Debs end up in /home/adoyle, one level above /home/adoyle/foo, as expected.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>