Pass arguments to docker build using --create --dockerarg 'args'
This creates a way to pass values to the 'docker build' command that
ultimately gets run to create the container. The specific instance
that demonstrated the need for this feature was the lack of a way to
pass Docker PROXY values to 'docker build'
So to achieve:
 docker build \
 --build-arg HTTP_PROXY=http://10.20.30.2:1234 \
 --build-arg FTP_PROXY=http://40.50.60.5:4567

One can now invoke DUE's --create  with
 --dockerarg "--build-arg HTTP_PROXY=http://10.20.30.2:1234"
 --dockerarg " --build-arg FTP_PROXY=http://40.50.60.5:4567"
OR
 --dockerarg "--build-arg HTTP_PROXY=http://10.20.30.2:1234 \
  --build-arg FTP_PROXY=http://40.50.60.5:4567"

Whatever follows --dockerarg is treated as a single argument, but
is passed to Docker unquoted for expansion.

Thanks to Billie Alsup at Cisco for pointing this out.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>