common-templates - container-create-user.sh Podman overhaul
Summary
-------
This fixes some glaring failures where the user did not retain
their identity when Podman was serving containers.

At this point the user should keep their user ID, group ID, and
thus their mounted home directories should be fully writeable,
regardless of how the containers are being run.

Some cosmetic formatting was also added to the configuration
output provided on container log in.

Longer explanation
------------------

The problem came out of Podman creating a user account in the
container, so that the container-create-user.sh script thought that
the account had already been configured, and did nothing.

In fixing that, some other behaviors have been updated as well.

So if the user does not already exist, they will be created.

Then configuration runs on the account that was just created, or
was already there, setting the shell, group ID's, sudo ability, etc.

Which should result in a user that can write to their home directory
as themselves and sudo in the container.

If the user is running a container using an account that is built
in to the container by specifiying --user-name, then the container
settings win, under the assumption that the internal configuration of
the container may count on the user ID/groups for that account
having certian values.
The user may want to specify a writeable --home-dir for this account as
due will still try to mount the invoking user's home directory.

Tests run
----------

Command:  due --run <select container>
 Home directory mounts.
 User id is preserved in container.
 User can sudo.
 Files created by user on host mounted home dir are owned by user.
 User can log in to running container.
 Files created by log in are user owned.

Command:  due --run --user-name <foo> <select container>
 User can log in to accounts already existing in container.
 Files created by existing accounts are owned by that account.
 That account cannot create files in the home mounted dir.

Command:  due --run --user-name <foo> --home-dir <world writeable system dir> <select container>
 User can run as user foo that already exists in the container, and
  can write files to it's home directory on the host supplied by --home-dir

Signed-off-by: Alex Doyle <adoyle@nvidia.com>