libdue - Reorder --login commands to print accurately
If --login is used to log in to a running container,
libdue now prints the actual command used for the log in,
as that could vary if the user was in a debug mode, or
on a Red Hat system.

Testing:
--------
Logged in to Fedora 33 and Debian contianers using Podman and
Docker, using an account in the container (--user-name) , and
had DUE create my account on the fly.
Also tested running commands like:
 ./due --run --command cd /bin \; ls -l

Signed-off-by: Alex Doyle <adoyle@nvidia.com>