Example post-install-config.sh.template - Add a user
One of the --login test cases is to log in using a user
account that already exists within the container.

This adds code to create a user account 'bob'.

It also illustrates some minor-but-fatal differences in how
Linux distributions can support particular commands, so while
the cases this code accounts for are not comprehensive, it
should provide a starting point for debugging future account
creation problems.

Note this code is _very_ similar to what container-create-user.sh
runs, so bug fixes added to one should probably be added to the other.

Testing:
--------
User account 'bob' was created in Debian 10 and Fedora 33 containers,
using Docker and Podman.

due --login --username bob
worked to log in to a running container as user bob.
due --run --username bob --home-dir <dir writeable by bob's UID>
Launched a container with the user logged in as bob, who could
save files in the directory.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>