container-create-user.sh - Shellcheck fixes and user creation cleanup
This applies a number of shellcheck cleanup suggestions,
and reformats the user creation code to use 'adduser' for both
Red Hat and Debian containers.

Since the versions of adduser differ between Fedora 33 and Debian 10,
some arguments are not supported in both cases and it has to be
conditional upon the distribution.

This should closely mirror the 'create a container with a user account'
code in templates/example/post-install-config.sh.template.

Testing:
--------
Used 'example' container because of its short creation time and
default user account 'bob'.

Tried the following on Debian 10 and Fedora 33 containers run by
both Docker and Podman

Could log in as myself and 'bob'

Ran:
 due --run --user-name <foo> --command whoami \; cd /due-configuration \; ls -l \; ls -l ~/

This shows that:
 Login works without crashing/throwing errors
 The user has the proper identity ( whoami )
 They can list a directory in the container ( context is correct )
 They can list their home directory ( username and groups should be correct )
 Multi argument command parsing is working correctly.

Signed-off-by: Alex Doyle <adoyle@nvidia.com>